python 3/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
numpy==1.22.3
protobuf==3.20.1
PyMySQL==1.0.2
pythainlp==3.0.8
requests==2.27.1
six==1.16.0
tinydb==4.7.0
torch==1.11.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
